fix: fix movie dto and add internet permission

+++ repository/src/main/java/com/cairosquad/repository/search/dataSource/remote/search/dto/MovieDto.kt
@@ -32,7 +32,7 @@ data class MovieDto(
     @SerialName("video")
     val video: Boolean?,
     @SerialName("vote_average")
-    val voteAverage: Int?,
+    val voteAverage: Double?,
     @SerialName("vote_count")
     val voteCount: Int?
 ) {
